backend/pyproject.toml
[project]
name = "django-insider"
version = "0.1.10"
description = "The missing observability dashboard for Django."
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    {name = "Morakinyo Joseph", email = "jmoraks03@gmail.com"}
]

keywords = ["django", "middleware", "monitoring", "performance", "celery", "react"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Django",
    "Framework :: Django :: 3.2",
    "Framework :: Django :: 4.0",
    "Framework :: Django :: 4.1",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "Django>=3.0,",
    "djangorestframework>=3.12",
    "celery>=5.0",
    "atlassian-python-api>=4.0",
    "requests>=2.30.0",
]

[project.urls]
Homepage = "https://github.com/Morakinyo-Joseph/django-insider"
Repository = "https://github.com/Morakinyo-Joseph/django-insider"
Documentation = "https://docs.google.com/document/d/1PnWshYnUwgQjzD5EDiG-HfkE37YIDSDdyvGJyhv2lM4/edit?tab=t.0"

[build-system]
requires = ["setuptools>=65.5.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["insider*"]
exclude = ["config*", "tests*", "venv*"]

[tool.setuptools.package-data]
insider = [
    "static/insider/index.html",
    "static/insider/assets/*",
    "templates/**/*",
    "migrations/**/*"
]
